Put new functions for loading JSON into configuration.cpp
