+++ src/HealthGPS.Console/jsonparser.h
@@ -2,7 +2,6 @@
 #include "options.h"
 #include "riskmodel.h"
 
-#include <filesystem>
 #include <nlohmann/json.hpp>
 #include <optional>
 
@@ -15,38 +14,6 @@ namespace host::poco {
 /// @sa https://github.com/nlohmann/json#arbitrary-types-conversions
 /// for details about the contents and code structure in this file.
 using json = nlohmann::json;
-
-/// @brief Get a path, based on base_dir, and check if it exists
-/// @param j Input JSON
-/// @param base_dir Base folder
-/// @return An absolute path, assuming that base_dir is the base if relative
-/// @throw json::type_error: Invalid JSON types
-/// @throw std::invalid_argument: Path does not exist
-std::filesystem::path get_valid_path(const json &j, const std::filesystem::path &base_dir);
-
-/// @brief Load FileInfo from JSON
-/// @param j Input JSON
-/// @param base_dir Base folder
-/// @return FileInfo
-/// @throw json::type_error: Invalid JSON types
-/// @throw std::invalid_argument: Path does not exist
-FileInfo get_file_info(const json &j, const std::filesystem::path &base_dir);
-
-/// @brief Load BaselineInfo from JSON
-/// @param j Input JSON
-/// @param base_dir Base folder
-/// @return BaselineInfo
-/// @throw json::type_error: Invalid JSON types
-/// @throw std::invalid_argument: Path does not exist
-BaselineInfo get_baseline_info(const json &j, const std::filesystem::path &base_dir);
-
-/// @brief Load ModellingInfo from JSON
-/// @param j Input JSON
-/// @param base_dir Base folder
-/// @return ModellingInfo
-/// @throw json::type_error: Invalid JSON types
-/// @throw std::invalid_argument: Path does not exist
-ModellingInfo get_modelling_info(const json &j, const std::filesystem::path &base_dir);
 
 //--------------------------------------------------------
 // Full risk factor model POCO types mapping
